==> scripts/06_predictive_model.qmd <==
---
title: "Middlebury placement assessment"
subtitle: "Fit predictive model"
author: "Joseph V. Casillas"
date: today
execute: 
  warning: false
  message: false
---

Load libraries, helper functions, and data. 

```{r}
#| label: setup
source(here::here("scripts", "05_load_data.R"))
```

```{r}
#| label: fig-lt-exam
#| fig-asp: 0.56
#| fig-dpi: 300
#| echo: false
mod_0 <- brm(
  formula = score_exam_z ~ score_lt_z, 
  chains = 4, cores = 4, threads = threading(2), 
  data = dat_2023, 
  file = here("models", "mod_0")
)

dat_2023 |> 
  ggplot() + 
  aes(x = score_lt_z, y = score_exam_z) + 
  geom_abline(
    data = slice_sample(as_tibble(mod_0), n = 200), 
    aes(intercept = b_Intercept, slope = b_score_lt_z), 
    alpha = 0.2, color = "grey40"
  ) + 
  geom_abline(
    data = summarize(as_tibble(mod_0), 
      intercept = mean(b_Intercept), slope = mean(b_score_lt_z)
    ), 
    aes(intercept = intercept, slope = slope), 
    color = "white", linewidth = 3
  ) + 
  geom_abline(
    data = summarize(as_tibble(mod_0), 
      intercept = mean(b_Intercept), slope = mean(b_score_lt_z)
    ), 
    aes(intercept = intercept, slope = slope), 
    color = "black", linewidth = 1.5
  ) + 
  geom_abline(intercept = 0, slope = 1, lty = 3) + 
  geom_point(
    aes(fill = level), 
    pch = 21, color = "white", size = 3, show.legend = F
  ) + 
  scale_fill_viridis_d(name = NULL) + 
  coord_cartesian(xlim = c(-2, 2), ylim = c(-2, 2)) + 
  ds4ling::ds4ling_bw_theme()
```


```{r}
#| label: fit-models
mod_null <- brm(
  formula = level_n ~ 1, 
  family = cumulative("probit"), 
  data = dat_2023, 
  chains = 4, cores = 4, threads = threading(2), 
  file = here("models", "mod_null")
)

mod_lt <- brm(
  formula = level_n ~ 1 + score_lt, 
  family = cumulative("probit"), 
  data = dat_2023, 
  chains = 4, cores = 4, threads = threading(2), 
  file = here("models", "mod_lt")
)

mod_placement <- brm(
  formula = level_n ~ 1 + score_exam, 
  family = cumulative("probit"), 
  data = dat_2023, 
  chains = 4, cores = 4, threads = threading(2), 
  file = here("models", "mod_placement")
)

mod_lt_placement <- brm(
  formula = level_n ~ 1 + score_lt + score_exam, 
  family = cumulative("probit"), 
  data = dat_2023, 
  chains = 4, cores = 4, threads = threading(2), 
  file = here("models", "mod_lt_placement")
)

```




```{r}
#| label: fig-model-pred-ex
#| fig-asp: 0.56
#| fig-dpi: 300
#| echo: false

pred_dat <- dat_2024 |> 
  filter(!is.na(score_exam) & !is.na(score_lt)) |> 
  select(id, score_exam, score_lt) |> 
  add_row(
    id = 11111, 
    score_exam = 20, 
    score_lt   = 40, 
  )

preds <- predictions(
  model = mod_lt_placement, 
  newdata = pred_dat, 
  type = "response"
) |> 
  posterior_draws() |> 
  mutate(
    id = as.factor(id), 
    group_lab = case_when(
      group == 1 ~ "1", 
      group == 2 ~ "1.5", 
      group == 3 ~ "2", 
      group == 4 ~ "3", 
      group == 5 ~ "4", 
      group == 6 ~ "Master"
    )
  ) 

# Summarize predictions for each ID that has taken the placement and LT
preds |> 
  group_by(id, group) |> 
  median_qi(draw) |> 
  mutate_if(is.numeric, specify_decimal, k = 2) |> 
  #mutate_if(is.numeric, formatC, format = 'f', flag = '0', digits = 2) |> 
  transmute(
    id, 
    draw, .lower, .upper, 
    group_lab = case_when(
      group == 1 ~ "1", 
      group == 2 ~ "1.5", 
      group == 3 ~ "2", 
      group == 4 ~ "3", 
      group == 5 ~ "4", 
      group == 6 ~ "Master"
    )
  ) |> 
  pivot_wider(
    names_from = "group_lab", 
    values_from = c("draw", ".lower", ".upper")
  ) |> 
  transmute(
    id, 
    level_1 = draw_1, 
    level_1_ci = glue("[{.lower_1}, {.upper_1}]"), 
    level_1.5 = draw_1.5, 
    level_1.5_ci = glue("[{.lower_1.5}, {.upper_1.5}]"), 
    level_2 = draw_2, 
    level_2_ci = glue("[{.lower_2}, {.upper_2}]"), 
    level_3 = draw_3, 
    level_3_ci = glue("[{.lower_3}, {.upper_3}]"), 
    level_4 = draw_4, 
    level_4_ci = glue("[{.lower_4}, {.upper_4}]"), 
    master = draw_Master, 
    master_ci = glue("[{.lower_Master}, {.upper_Master}]") 
  ) |> 
  saveRDS(here("data", "tidy", "preds_2024.rds"))


medians <- preds |> 
  filter(id == "11111") |> 
  group_by(group_lab) |> 
  mean_qi(draw)

preds |> 
  filter(id == "11111") |> 
  ggplot() + 
  aes(x = group_lab, y = draw) + 
  geom_segment(
    data = medians, 
    aes(x = group_lab, y = .lower, yend = .upper), 
    linewidth = 3, lineend = "round", color = "grey60"
  ) + 
  geom_point(
    data = medians, 
    aes(x = group_lab, y = draw), 
    pch = 21, stroke = 2, size = 3
  ) + 
  coord_cartesian(ylim = c(0, 1)) + 
  labs(x = NULL, y = "p(level)") + 
  ds4ling::ds4ling_bw_theme()

```
